refactor bamBam and extract command exacution from cli, catch errors at cli level
